+++ DatadogInternal/Sources/Models/SessionReplay/SessionReplayConfiguration.swift
@@ -8,6 +8,8 @@ import Foundation
 
 public let SessionReplayFeaturneName = "session-replay"
 
+// MARK: Deprecated Global Privacy Level
+
 /// Available privacy levels for content masking in Session Replay.
 public enum SessionReplayPrivacyLevel: String {
     /// Record all content.
@@ -17,8 +19,45 @@ public enum SessionReplayPrivacyLevel: String {
     case mask
 
     /// Mask input elements, but record all other content.
-    case maskUserInput = "mask_user_input"
+    case maskUserInput = "mask-user-input"
 }
+
+// MARK: Fine-Grained Privacy Levels
+
+/// Available privacy levels for text and input masking in Session Replay.
+public enum TextAndInputPrivacyLevel: String, CaseIterable {
+    /// Show all texts except sensitive inputs, eg. password fields.
+    case maskSensitiveInputs = "mask_sensitive_inputs"
+
+    /// Mask all inputs fields, eg. textfields, switches, checkboxes.
+    case maskAllInputs = "mask_all_inputs"
+
+    /// Mask all texts and inputs, eg. labels.
+    case maskAll = "mask_all"
+}
+
+/// Available privacy levels for image masking in the Session Replay.
+public enum ImagePrivacyLevel: String {
+    /// Only SF Symbols and images loaded using UIImage(named:) that are bundled within the application will be recorded.
+    case maskNonBundledOnly = "mask_non_bundled_only"
+
+    /// No images will be recorded.
+    case maskAll = "mask_all"
+
+    /// All images will be recorded, including the ones downloaded from the Internet or generated during the app runtime.
+    case maskNone = "mask_none"
+}
+
+/// Available privacy levels for touch masking in Session Replay.
+public enum TouchPrivacyLevel: String {
+    /// Show all user touches.
+    case show
+
+    /// Hide all user touches.
+    case hide
+}
+
+// MARK: SessionReplayConfiguration
 
 /// The Session Replay shared configuration.
 ///
@@ -32,8 +71,10 @@ public enum SessionReplayPrivacyLevel: String {
 ///     )
 ///
 public protocol SessionReplayConfiguration {
-    /// The privacy level to use for the web view replay recording.
-    var privacyLevel: SessionReplayPrivacyLevel { get }
+    /// Fine-Grained privacy levels to use in Session Replay.
+    var textAndInputPrivacyLevel: TextAndInputPrivacyLevel { get }
+    var imagePrivacyLevel: ImagePrivacyLevel { get }
+    var touchPrivacyLevel: TouchPrivacyLevel { get }
 }
 
 extension DatadogFeature where Self: SessionReplayConfiguration {