+++ DatadogSessionReplay/SRSnapshotTests/SRSnapshotTests/Utils/SnapshotTestCase.swift
@@ -11,8 +11,10 @@ import TestUtilities
 @testable import DatadogSessionReplay
 @testable import SRHost
 
-private var defaultPrivacyLevel: SessionReplayPrivacyLevel {
-    return SessionReplay.Configuration(replaySampleRate: 100).defaultPrivacyLevel
+private var defaultImagePrivacyLevel: ImagePrivacyLevel = .maskNonBundledOnly
+private var defaultTextAndInputPrivacyLevel: TextAndInputPrivacyLevel = .maskSensitiveInputs
+internal var allTextAndInputPrivacyLevels: [TextAndInputPrivacyLevel] {
+    return TextAndInputPrivacyLevel.allCases
 }
 
 internal class SnapshotTestCase: XCTestCase {
@@ -38,7 +40,8 @@ internal class SnapshotTestCase: XCTestCase {
     /// Helper method for most snapshot tests
     func takeSnapshotFor(
         _ fixture: Fixture,
-        with privacyModes: [SessionReplayPrivacyLevel] = [defaultPrivacyLevel],
+        with textAndInputPrivacyLevels: [TextAndInputPrivacyLevel] = [defaultTextAndInputPrivacyLevel],
+        imagePrivacyLevel: ImagePrivacyLevel = defaultImagePrivacyLevel,
         shouldRecord: Bool,
         folderPath: String,
         fileNamePrefix: String? = nil,
@@ -49,9 +52,9 @@ internal class SnapshotTestCase: XCTestCase {
         // Give time for the view to appear and lay out properly
         wait(seconds: 0.2)
 
-        try forPrivacyModes(privacyModes) { privacyMode in
-            let image = try takeSnapshot(with: privacyMode)
-            let fileNameSuffix = fileNamePrefix == nil ? "-\(privacyMode)-privacy" : "-\(fileNamePrefix!)-\(privacyMode)-privacy"
+        try forPrivacyModes(textAndInputPrivacyLevels) { textPrivacyLevel in
+            let image = try takeSnapshot(with: textPrivacyLevel, imagePrivacyLevel: imagePrivacyLevel)
+            let fileNameSuffix = fileNamePrefix == nil ? "-\(textPrivacyLevel)-privacy" : "-\(fileNamePrefix!)-\(textPrivacyLevel)-privacy"
             let snapshotLocation: ImageLocation = .folder(named: folderPath, fileNameSuffix: fileNameSuffix, file: file, function: function)
 
             DDAssertSnapshotTest(
@@ -93,7 +96,7 @@ internal class SnapshotTestCase: XCTestCase {
         fixture: Fixture,
         showPopup: (PopupsViewController) -> Void,
         waitTime: TimeInterval,
-        privacyModes: [SessionReplayPrivacyLevel] = [defaultPrivacyLevel],
+        textPrivacyModes: [TextAndInputPrivacyLevel] = [defaultTextAndInputPrivacyLevel],
         shouldRecord: Bool,
         folderPath: String,
         file: StaticString = #filePath,
@@ -102,20 +105,22 @@ internal class SnapshotTestCase: XCTestCase {
         let vc = show(fixture: fixture) as! PopupsViewController
         showPopup(vc)
         wait(seconds: waitTime)
-        print("privacyModes:", privacyModes)
 
-        for privacyMode in privacyModes {
-            let image = try takeSnapshot(with: privacyMode)
+        for textPrivacyMode in textPrivacyModes {
+            let image = try takeSnapshot(with: textPrivacyMode)
             DDAssertSnapshotTest(
                 newImage: image,
-                snapshotLocation: .folder(named: folderPath, fileNameSuffix: "-\(privacyMode)-privacy", file: file, function: function),
+                snapshotLocation: .folder(named: folderPath, fileNameSuffix: "-\(textPrivacyMode)-privacy", file: file, function: function),
                 record: shouldRecord
             )
         }
     }
 
     /// Captures side-by-side snapshot of the app UI and recorded wireframes.
-    func takeSnapshot(with privacyLevel: SessionReplayPrivacyLevel = defaultPrivacyLevel) throws -> UIImage {
+    func takeSnapshot(
+        with textAndInputPrivacyLevel: TextAndInputPrivacyLevel = defaultTextAndInputPrivacyLevel,
+        imagePrivacyLevel: ImagePrivacyLevel = defaultImagePrivacyLevel
+    ) throws -> UIImage {
         let expectWireframes = self.expectation(description: "Wait for wireframes")
         let expectResources = self.expectation(description: "Wait for resources")
 
@@ -154,7 +159,15 @@ internal class SnapshotTestCase: XCTestCase {
 
         // Capture next record with mock RUM Context
         try recorder.captureNextRecord(
-            .init(privacy: privacyLevel, applicationID: "", sessionID: "", viewID: "", viewServerTimeOffset: 0)
+            .init(
+                textAndInputPrivacy: textAndInputPrivacyLevel,
+                imagePrivacy: imagePrivacyLevel,
+                touchPrivacy: .show,
+                applicationID: "",
+                sessionID: "",
+                viewID: "",
+                viewServerTimeOffset: 0
+            )
         )
 
         waitForExpectations(timeout: 30) // very pessimistic timeout to mitigate CI lags
@@ -170,13 +183,13 @@ internal class SnapshotTestCase: XCTestCase {
         // Add XCTest attachements for debugging and troubleshooting:
         // - attach recorded wireframes as JSON
         let wireframesAttachement = XCTAttachment(string: renderedWireframes.debugInfo.dumpWireframesAsJSON())
-        wireframesAttachement.name = "recorded-wireframes-(\(privacyLevel)).json"
+        wireframesAttachement.name = "recorded-wireframes-(\(textAndInputPrivacyLevel)).json"
         wireframesAttachement.lifetime = .deleteOnSuccess
         add(wireframesAttachement)
 
         // - attach rendered as blueprint text
         let blueprintAttachement = XCTAttachment(string: renderedWireframes.debugInfo.dumpImageAsBlueprint())
-        blueprintAttachement.name = "rendered-blueprint-(\(privacyLevel)).txt"
+        blueprintAttachement.name = "rendered-blueprint-(\(textAndInputPrivacyLevel)).txt"
         blueprintAttachement.lifetime = .deleteOnSuccess
         add(blueprintAttachement)
 
@@ -195,8 +208,8 @@ internal class SnapshotTestCase: XCTestCase {
     }
 
     func forPrivacyModes(
-        _ modes: [SessionReplayPrivacyLevel] = [.mask, .allow, .maskUserInput],
-        do work: (SessionReplayPrivacyLevel) throws -> Void) rethrows {
+        _ modes: [TextAndInputPrivacyLevel] = TextAndInputPrivacyLevel.allCases,
+        do work: (TextAndInputPrivacyLevel) throws -> Void) rethrows {
         try modes.forEach { try work($0) }
     }
 }