+++ DatadogSessionReplay/Sources/Recorder/RecordingCoordinator.swift
@@ -16,7 +16,9 @@ internal class RecordingCoordinator {
     let recorder: Recording
     let scheduler: Scheduler
     let sampler: Sampler
-    let privacy: PrivacyLevel
+    let textAndInputPrivacy: TextAndInputPrivacyLevel
+    let imagePrivacy: ImagePrivacyLevel
+    let touchPrivacy: TouchPrivacyLevel
     let srContextPublisher: SRContextPublisher
 
     private var currentRUMContext: RUMContext? = nil
@@ -33,7 +35,9 @@ internal class RecordingCoordinator {
 
     init(
         scheduler: Scheduler,
-        privacy: PrivacyLevel,
+        textAndInputPrivacy: TextAndInputPrivacyLevel,
+        imagePrivacy: ImagePrivacyLevel,
+        touchPrivacy: TouchPrivacyLevel,
         rumContextObserver: RUMContextObserver,
         srContextPublisher: SRContextPublisher,
         recorder: Recording,
@@ -45,7 +49,9 @@ internal class RecordingCoordinator {
         self.recorder = recorder
         self.scheduler = scheduler
         self.sampler = sampler
-        self.privacy = privacy
+        self.textAndInputPrivacy = textAndInputPrivacy
+        self.imagePrivacy = imagePrivacy
+        self.touchPrivacy = touchPrivacy
         self.srContextPublisher = srContextPublisher
         self.telemetry = telemetry
         self.methodCallTelemetrySamplingRate = methodCallTelemetrySamplingRate
@@ -118,7 +124,9 @@ internal class RecordingCoordinator {
         }
 
         let recorderContext = Recorder.Context(
-            privacy: privacy,
+            textAndInputPrivacy: textAndInputPrivacy,
+            imagePrivacy: imagePrivacy,
+            touchPrivacy: touchPrivacy,
             applicationID: rumContext.applicationID,
             sessionID: rumContext.sessionID,
             viewID: viewID,