+++ DatadogSessionReplay/Sources/SessionReplayConfiguration.swift
@@ -10,6 +10,9 @@ import DatadogInternal
 
 // swiftlint:disable duplicate_imports
 @_exported import enum DatadogInternal.SessionReplayPrivacyLevel
+@_exported import enum DatadogInternal.TextAndInputPrivacyLevel
+@_exported import enum DatadogInternal.ImagePrivacyLevel
+@_exported import enum DatadogInternal.TouchPrivacyLevel
 // swiftlint:enable duplicate_imports
 
 extension SessionReplay {
@@ -28,7 +31,31 @@ extension SessionReplay {
         /// Defines the way sensitive content (e.g. text) should be masked.
         ///
         /// Default: `.mask`.
-        public var defaultPrivacyLevel: SessionReplayPrivacyLevel
+        @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use the new privacy levels instead.")
+        public var defaultPrivacyLevel: SessionReplayPrivacyLevel = .mask {
+            /// Whenever a new `defaultPrivacyLevel` is set, it converts it to the new privacy levels.
+            didSet {
+                let newPrivacyLevels = Self.convertPrivacyLevel(from: defaultPrivacyLevel)
+                self.textAndInputPrivacyLevel = newPrivacyLevels.textAndInputPrivacy
+                self.imagePrivacyLevel = newPrivacyLevels.imagePrivacy
+                self.touchPrivacyLevel = newPrivacyLevels.touchPrivacy
+            }
+        }
+
+        /// Defines the way text and input (e.g. textfields, checkboxes) should be masked.
+        ///
+        /// Default: `.maskAll`.
+        public var textAndInputPrivacyLevel: TextAndInputPrivacyLevel
+
+        /// Defines image privacy level.
+        ///
+        /// Default: `.maskAll`.
+        public var imagePrivacyLevel: ImagePrivacyLevel
+
+        /// Defines the way user touches (e.g. tap) should be masked.
+        ///
+        /// Default: `.hide`.
+        public var touchPrivacyLevel: TouchPrivacyLevel
 
         /// Defines it the recording should start automatically. When `true`, the recording starts automatically; when `false` it doesn't, and the recording will need to be started manually.
         ///
@@ -49,9 +76,34 @@ extension SessionReplay {
         /// Creates Session Replay configuration
         /// - Parameters:
         ///   - replaySampleRate: The sampling rate for Session Replay. It is applied in addition to the RUM session sample rate.
+        ///   - textAndInputPrivacyLevel: The way texts and inputs (e.g. label, textfield, checkbox) should be masked. Default: `.maskAll`.
+        ///   - touchPrivacyLevel: The way user touches (e.g. tap) should be masked. Default: `.hide`.
+        ///   - defaultImageRecordingLevel: Image recording privacy level. Default: `.maskAll`.
+
+        ///   - customEndpoint: Custom server url for sending replay data. Default: `nil`.
+        public init(
+            replaySampleRate: Float,
+            textAndInputPrivacyLevel: TextAndInputPrivacyLevel,
+            imagePrivacyLevel: ImagePrivacyLevel,
+            touchPrivacyLevel: TouchPrivacyLevel,
+            startRecordingImmediately: Bool = true,
+            customEndpoint: URL? = nil
+        ) {
+            self.replaySampleRate = replaySampleRate
+            self.textAndInputPrivacyLevel = textAndInputPrivacyLevel
+            self.imagePrivacyLevel = imagePrivacyLevel
+            self.touchPrivacyLevel = touchPrivacyLevel
+            self.startRecordingImmediately = startRecordingImmediately
+            self.customEndpoint = customEndpoint
+        }
+
+        /// Creates Session Replay configuration.
+        /// - Parameters:
+        ///   - replaySampleRate: The sampling rate for Session Replay. It is applied in addition to the RUM session sample rate.
         ///   - defaultPrivacyLevel: The way sensitive content (e.g. text) should be masked. Default: `.mask`.
         ///   - startRecordingImmediately: If the recording should start automatically. When `true`, the recording starts automatically; when `false` it doesn't, and the recording will need to be started manually. Default: `true`.
         ///   - customEndpoint: Custom server url for sending replay data. Default: `nil`.
+        @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use `init(replaySampleRate:textAndInputPrivacyLevel:imagePrivacyLevel:touchPrivacyLevel:)` instead.")
         public init(
             replaySampleRate: Float,
             defaultPrivacyLevel: SessionReplayPrivacyLevel = .mask,
@@ -60,6 +112,10 @@ extension SessionReplay {
         ) {
             self.replaySampleRate = replaySampleRate
             self.defaultPrivacyLevel = defaultPrivacyLevel
+            let newPrivacyLevels = Self.convertPrivacyLevel(from: defaultPrivacyLevel)
+            self.textAndInputPrivacyLevel = newPrivacyLevels.textAndInputPrivacy
+            self.imagePrivacyLevel = newPrivacyLevels.imagePrivacy
+            self.touchPrivacyLevel = newPrivacyLevels.touchPrivacy
             self.startRecordingImmediately = startRecordingImmediately
             self.customEndpoint = customEndpoint
         }
@@ -67,6 +123,35 @@ extension SessionReplay {
         @_spi(Internal)
         public mutating func setAdditionalNodeRecorders(_ additionalNodeRecorders: [SessionReplayNodeRecorder]) {
             self._additionalNodeRecorders = additionalNodeRecorders
+        }
+
+        /// Method to convert deprecated `SessionReplayPrivacyLevel` to the new privacy levels.
+        internal static func convertPrivacyLevel(from oldPrivacyLevel: SessionReplayPrivacyLevel)
+        -> (
+            textAndInputPrivacy: TextAndInputPrivacyLevel,
+            imagePrivacy: ImagePrivacyLevel,
+            touchPrivacy: TouchPrivacyLevel
+        ) {
+            switch oldPrivacyLevel {
+            case .allow:
+                return (
+                    textAndInputPrivacy: .maskSensitiveInputs,
+                    imagePrivacy: .maskNone,
+                    touchPrivacy: .show
+                )
+            case .maskUserInput:
+                return (
+                    textAndInputPrivacy: .maskAllInputs,
+                    imagePrivacy: .maskNonBundledOnly,
+                    touchPrivacy: .hide
+                )
+            case .mask:
+                return (
+                    textAndInputPrivacy: .maskAll,
+                    imagePrivacy: .maskAll,
+                    touchPrivacy: .hide
+                )
+            }
         }
     }
 }