+++ DatadogSessionReplay/Tests/Feature/RequestBuilder/JSON/SegmentJSONTests.swift
@@ -144,7 +144,9 @@ class SegmentJSONTests: XCTestCase {
 
     private func generateEnrichedRecordJSONs(for segment: SRSegment) throws -> [SegmentJSON] {
         let context = Recorder.Context(
-            privacy: .mockRandom(),
+            textAndInputPrivacy: .mockRandom(),
+            imagePrivacy: .mockRandom(),
+            touchPrivacy: .mockRandom(),
             rumContext: RUMContext(
                 applicationID: segment.application.id,
                 sessionID: segment.session.id,