+++ DatadogSessionReplay/Tests/Processor/SnapshotProcessorTests.swift
@@ -235,7 +235,13 @@ class SnapshotProcessorTests: XCTestCase {
         // When
         let snapshot = ViewTreeSnapshot(
             date: time,
-            context: .init(privacy: .allow, rumContext: rum, date: time),
+            context: .init(
+                textAndInputPrivacy: .mockRandom(),
+                imagePrivacy: .mockRandom(),
+                touchPrivacy: .mockRandom(),
+                rumContext: rum,
+                date: time
+            ),
             viewportSize: .mockRandom(minWidth: 1_000, minHeight: 1_000),
             nodes: [node],
             webViewSlotIDs: Set([hiddenSlot, visibleSlot])
@@ -306,7 +312,8 @@ class SnapshotProcessorTests: XCTestCase {
             attributes: .mockAny(),
             contentFrame: .mockAny(),
             clipsToBounds: .mockAny(),
-            imageResource: resource
+            imageResource: resource,
+            imagePrivacyLevel: .maskNonBundledOnly
         )
         let snapshot: ViewTreeSnapshot = .mockWith(
             context: .mockRandom(),
@@ -436,7 +443,16 @@ class SnapshotProcessorTests: XCTestCase {
     private let snapshotBuilder = ViewTreeSnapshotBuilder(additionalNodeRecorders: [])
 
     private func generateViewTreeSnapshot(for viewTree: UIView, date: Date, rumContext: RUMContext) -> ViewTreeSnapshot {
-        snapshotBuilder.createSnapshot(of: viewTree, with: .init(privacy: .allow, rumContext: rumContext, date: date))
+        snapshotBuilder.createSnapshot(
+            of: viewTree,
+            with: .init(
+                textAndInputPrivacy: .mockRandom(),
+                imagePrivacy: .mockRandom(),
+                touchPrivacy: .mockRandom(),
+                rumContext: rumContext,
+                date: date
+            )
+        )
     }
 
     private func generateSimpleViewTree() -> UIView {