+++ DatadogSessionReplay/Tests/SessionReplayTests.swift
@@ -63,35 +63,52 @@ class SessionReplayTests: XCTestCase {
         // Then
         let sr = try XCTUnwrap(core.get(feature: SessionReplayFeature.self))
         XCTAssertEqual(sr.recordingCoordinator.sampler.samplingRate, 42)
-        XCTAssertEqual(sr.recordingCoordinator.privacy, .mask)
+        XCTAssertEqual(sr.recordingCoordinator.textAndInputPrivacy, .maskAll)
+        XCTAssertEqual(sr.recordingCoordinator.imagePrivacy, .maskAll)
+        XCTAssertEqual(sr.recordingCoordinator.touchPrivacy, .hide)
         XCTAssertNil((sr.requestBuilder as? SegmentRequestBuilder)?.customUploadURL)
         let r = try XCTUnwrap(core.get(feature: ResourcesFeature.self))
         XCTAssertNil((r.requestBuilder as? ResourceRequestBuilder)?.customUploadURL)
     }
 
-    func testWhenEnabled_itSendsConfigurationTelemetry() throws {
-        // Given
-        let sampleRate: Int64 = .mockRandom(min: 0, max: 100)
-        let privacyLevel: SessionReplayPrivacyLevel = .mockRandom()
-        let startRecordingImmediately: Bool = .mockRandom()
-        let messageReceiver = FeatureMessageReceiverMock()
-        let core = PassthroughCoreMock(messageReceiver: messageReceiver)
-
-        // When
-        SessionReplay.enable(
-            with: SessionReplay.Configuration(
-                replaySampleRate: Float(sampleRate),
-                defaultPrivacyLevel: privacyLevel,
-                startRecordingImmediately: startRecordingImmediately
-            ),
-            in: core
+    func testWhenEnabledWithNewAPI() throws {
+        let textAndInputPrivacy: TextAndInputPrivacyLevel = .mockRandom()
+        let imagePrivacy: ImagePrivacyLevel = .mockRandom()
+        let touchPrivacy: TouchPrivacyLevel = .mockRandom()
+        config = SessionReplay.Configuration(
+            replaySampleRate: 42,
+            textAndInputPrivacyLevel: textAndInputPrivacy,
+            imagePrivacyLevel: imagePrivacy,
+            touchPrivacyLevel: touchPrivacy
         )
 
+        // When
+        SessionReplay.enable(with: config, in: core)
+
         // Then
-        let configuration = try XCTUnwrap(messageReceiver.messages.firstTelemetry?.asConfiguration)
-        XCTAssertEqual(configuration.sessionReplaySampleRate, sampleRate)
-        XCTAssertEqual(configuration.defaultPrivacyLevel, privacyLevel.rawValue)
-        XCTAssertEqual(configuration.startRecordingImmediately, startRecordingImmediately)
+        let sr = try XCTUnwrap(core.get(feature: SessionReplayFeature.self))
+        XCTAssertEqual(sr.recordingCoordinator.sampler.samplingRate, 42)
+        XCTAssertEqual(sr.recordingCoordinator.textAndInputPrivacy, textAndInputPrivacy)
+        XCTAssertEqual(sr.recordingCoordinator.imagePrivacy, imagePrivacy)
+        XCTAssertEqual(sr.recordingCoordinator.touchPrivacy, touchPrivacy)
+        XCTAssertNil((sr.requestBuilder as? SegmentRequestBuilder)?.customUploadURL)
+        let r = try XCTUnwrap(core.get(feature: ResourcesFeature.self))
+        XCTAssertNil((r.requestBuilder as? ResourceRequestBuilder)?.customUploadURL)
+    }
+
+    func testWhenEnabledWithRandomPrivacyLevel() throws {
+        let randomPrivacy: SessionReplayPrivacyLevel = .mockRandom()
+        config.defaultPrivacyLevel = randomPrivacy
+
+        // When
+        SessionReplay.enable(with: config, in: core)
+
+        // Then
+        let sr = try XCTUnwrap(core.get(feature: SessionReplayFeature.self))
+        let newPrivacyLevels = SessionReplay.Configuration.convertPrivacyLevel(from: randomPrivacy)
+        XCTAssertEqual(sr.recordingCoordinator.textAndInputPrivacy, newPrivacyLevels.textAndInputPrivacy)
+        XCTAssertEqual(sr.recordingCoordinator.imagePrivacy, newPrivacyLevels.imagePrivacy)
+        XCTAssertEqual(sr.recordingCoordinator.touchPrivacy, newPrivacyLevels.touchPrivacy)
     }
 
     func testWhenEnabledWithReplaySampleRate() throws {
@@ -104,18 +121,6 @@ class SessionReplayTests: XCTestCase {
         // Then
         let sr = try XCTUnwrap(core.get(feature: SessionReplayFeature.self))
         XCTAssertEqual(sr.recordingCoordinator.sampler.samplingRate, random)
-    }
-
-    func testWhenEnabledWithDefaultPrivacyLevel() throws {
-        let random: PrivacyLevel = .mockRandom()
-        config.defaultPrivacyLevel = random
-
-        // When
-        SessionReplay.enable(with: config, in: core)
-
-        // Then
-        let sr = try XCTUnwrap(core.get(feature: SessionReplayFeature.self))
-        XCTAssertEqual(sr.recordingCoordinator.privacy, random)
     }
 
     func testWhenEnabledWithCustomEndpoint() throws {
@@ -167,6 +172,69 @@ class SessionReplayTests: XCTestCase {
         // Then
         XCTAssertNil(core.get(feature: SessionReplayFeature.self))
         XCTAssertNil(core.get(feature: ResourcesFeature.self))
+    }
+
+    // MARK: Telemetry
+
+    func testWhenEnabled_itSendsConfigurationTelemetry() throws {
+        // Given
+        let sampleRate: Int64 = .mockRandom(min: 0, max: 100)
+        let privacyLevel: SessionReplayPrivacyLevel = .mockRandom()
+        let startRecordingImmediately: Bool = .mockRandom()
+        let messageReceiver = FeatureMessageReceiverMock()
+        let core = PassthroughCoreMock(messageReceiver: messageReceiver)
+
+        // When
+        SessionReplay.enable(
+            with: SessionReplay.Configuration(
+                replaySampleRate: Float(sampleRate),
+                defaultPrivacyLevel: privacyLevel,
+                startRecordingImmediately: startRecordingImmediately
+            ),
+            in: core
+        )
+
+        // Then
+        let configuration = try XCTUnwrap(messageReceiver.messages.firstTelemetry?.asConfiguration)
+        XCTAssertEqual(configuration.sessionReplaySampleRate, sampleRate)
+        XCTAssertNil(configuration.defaultPrivacyLevel)
+        let newPrivacyLevels = SessionReplay.Configuration.convertPrivacyLevel(from: privacyLevel)
+        XCTAssertEqual(configuration.textAndInputPrivacyLevel, newPrivacyLevels.textAndInputPrivacy.rawValue)
+        XCTAssertEqual(configuration.imagePrivacyLevel, newPrivacyLevels.imagePrivacy.rawValue)
+        XCTAssertEqual(configuration.touchPrivacyLevel, newPrivacyLevels.touchPrivacy.rawValue)
+        XCTAssertEqual(configuration.startRecordingImmediately, startRecordingImmediately)
+    }
+
+    func testWhenEnabled_itSendsConfigurationTelemetry_withNewApi() throws {
+        // Given
+        let sampleRate: Int64 = .mockRandom(min: 0, max: 100)
+        let textAndInputLevel: TextAndInputPrivacyLevel = .mockRandom()
+        let imageLevel: ImagePrivacyLevel = .mockRandom()
+        let touchLevel: TouchPrivacyLevel = .mockRandom()
+        let startRecordingImmediately: Bool = .mockRandom()
+        let messageReceiver = FeatureMessageReceiverMock()
+        let core = PassthroughCoreMock(messageReceiver: messageReceiver)
+
+        // When
+        SessionReplay.enable(
+            with: SessionReplay.Configuration(
+                replaySampleRate: Float(sampleRate),
+                textAndInputPrivacyLevel: textAndInputLevel,
+                imagePrivacyLevel: imageLevel,
+                touchPrivacyLevel: touchLevel,
+                startRecordingImmediately: startRecordingImmediately
+            ),
+            in: core
+        )
+
+        // Then
+        let configuration = try XCTUnwrap(messageReceiver.messages.firstTelemetry?.asConfiguration)
+        XCTAssertEqual(configuration.sessionReplaySampleRate, sampleRate)
+        XCTAssertNil(configuration.defaultPrivacyLevel)
+        XCTAssertEqual(configuration.textAndInputPrivacyLevel, textAndInputLevel.rawValue)
+        XCTAssertEqual(configuration.imagePrivacyLevel, imageLevel.rawValue)
+        XCTAssertEqual(configuration.touchPrivacyLevel, touchLevel.rawValue)
+        XCTAssertEqual(configuration.startRecordingImmediately, startRecordingImmediately)
     }
 
     // MARK: - Recording Tests